New Tests
Add more tests. The Images are not included as I currently do not have
the theming available and so the images will lead to failures in a
themed environment.

All test at least do not throw any exception, but the images do not
look good in most case and should be checked if the tests make sense
or the correspondig code needs improvements.