split __repr__ into draw and __repr__ and set interactive=False in draw()
`draw()` only draws the plot without showing it, `__repr__()` then uses
plt.show() to open the plot (which only does it if interactive == True
and a suitable backend is used). This change is needed for saving a
plot via `ggsave(plot)`.

To make this possible, interactive = True has to be set to False
while in `draw()`. Set interactive=True for the rest
(= `__repr__()`) during importing and based on `sys.flags.interactive`.