FIX: Check for x/y == None in facet_wrap and new tests
If you only supply one variable (x or y) to facet_wrap, the other
would be "None", resulting in a `df.get(None)` call, which doesn't
anymore work in newer pandas. So add a few checks for None.

Also change the __radd__() method, so that only the copy of the ggplot
object is changed. This also needs adjustments in the tests.

Decouple visual tests and tests of internals in test_scale_facet_wrap

Add a new test that __radd__() does not change the original object.