When geom aesthetics become stat parameters
In geom initialization, any argument that is both an aesthetic to
to the geom and a parameter setting to the stat is taken by the
stat if it is not an array-like. This is useful where the weight is
an aestetic for a geom, yet it is used during stat computation.
For this case, the problem was the API value for the weight is 1
(a scalar), yet it is also mappable. Although scalar weights are
essentially meaningless they do manipulate the y-scaling of histograms.

geoms abline, vline and hline can take functions to compute the slope
and intercepts. The functions are used by the respective stats.