Fix issue with color brewer
Whenever we don't have a color specified or use 'fill' but not color,
colorbrewer will throw an exception because it can't figure out how many
colors we are using. Particularly for the case of 'fill' this is a
problem since bar plots for example often will specify 'fill' without
specifying 'color'.

Now color brewer checks for 'fill' as well and if it finds neither
'fill' nor color it just asks for 3 colors from the color theme (any
less will cause an exception).