No more dataframe.groupby().apply()
Pandas calls the function two times on the first grouped dataframe.
This was making it hard to debug, and in cases where the function
calls code that also does a `dataframe.groupby().apply` the
extra calls multiply.

The solution is to do this with a `groupby_apply` utility
function that splits-applys-combines in a straightforward way.