Fixed issue for data frames with no 0 index
Saving a dataframe with no 0-indices results in a KeyError. This is
because calling `[0]` on a dataframe doesn't return the first row but instead
the row assigned to 0-index. If no 0-index exists then using `[0]`
throws a KeyError. For example:

```python
df = pd.DataFrame({ 999 : { "test": 0 }})
df[0] # Results in KeyError
df.iloc[0] # Returns the proper line with index 999
```

The usage of `[0]` caused ggsave to throw a KeyError exception when
saving dataframes without a 0-index. The usage of `[0]` has been
changed to `.iloc[0]` to fix this issue.