Rebased and seem successful
This is the first commit after a big rebase using the command

`git rebase --strategy=recursive --strategy-option=theirs upstream/master`

The effect of this is that `upstream/master` is clobbered in case of
merge conflicts and only new files have survived. The branches had
diverged at commit 29ec5abad4b620f575c12f96eef914595524f28d. With the
rebase overlaying a near rewrite not much has been lost in terms
functionality, however that related to the new files that have come in
is going to be integerated. i.e the new color scheme and geoms. Although
not verified, it is very likely that some new test cases have not survived.

If you are tracking the changes and are reading this, I am remorseful
for any undue suffering that my actions have caused.

In this commit obsolete import statements have been removed, this gets
everything functional.