Do not use formatter offset for the guide
Unlike MPL axes for which very small or very large numbers
may be displayed using an offset e.g(10e-7), the ggplot guide
does not do that. We have to tell the formatter not try and
be clever in case small or large numbers show up. Otherwise
the wrong values get displayed.