Fix size aesthetic scale factor
For points, the size as specified by the user or
returned by any of the size scales is now multiplied
by `pi`. This gives a better perception for mappings
with `scale_size_continuous` and `scale_size_area`.
As MPL scatter/polygons are specified in `points^2`
this means that our points are in `pi*points^2`. The
results are also closer to ggplot2.

For lines, the size is now multiplied by `sqrt(pi)`.
This is because MPL lines are specified in `points`,
but we want, for the same size, the diameter of
circular points to match the linewidth of the lines.
This means our lines sizes are in `sqrt(pi)*points`,
because of the `pi` factor on the points.