Refactor themeables, add setup_figure
Some `apply`s are now `apply_figure` as they should have
been, since they modify the figure and need to so once.

The distinction betweeen `setup_figure` and `apply_figure`
is necessary because some `figure` params (especially those
that affect the spacing) will influence the placement of
objects (legend) on the plot.