Add element_bank(), implement blanking & refactor
This commit adds full support for creating themes using
only theme elements and no `rcParams`.

It also fixes how themes are added so that inheritance
hierarchy of the themeables is taken into account. This
prevents ensures that when `theme_second` is added to
`theme_first` that parent themeables in `theme_second`
mix into or overide the child elements.

Still appreciating @gdowding for that `mro` based
forward inheritance trick.