Fix multiple panels with layer specific data
For a faceted plot with many layers, if a layer did not have data
for a given panel this would error out, for facet_wrap.

The solution is not to limit the(unsed) panel categories
(facet_grid was doing this) but to detect the empty panel data
splitting and combining then bailing out appropriately.