Move examples to a separate package
*Reason*

The notebooks should be stored in executed form with the output
cells so that they are viewable online. The output will largely
consist of images that would change when the notebooks are
re-executed. This would lead to a bloated repository.

The solution is a separate repository (that can be rebased with
limited consequences) that is a python package which can be
installed by the document building system.

Using a submodule was not appealing.