Do not draw ggplot when it has a figure
The internals are not clean enough to allow a user
to draw a ggplot object twice. It does not come up
in everyday use, but it may arise when using the
internal functions that PlotAnimation uses. Even
then it is only likely to show up when working in
a Jupyter Notebook.

Helps people doing complicated stuff like
https://github.com/has2k1/plotnine/issues/68#issuecomment-347674619