Fix failing formula without dot in facet_grid
The facet_grid was failing if a dot was not specified for single
row/colum formulas. See test case
test_facet_grid_scales_free_y_formula_no_dot. It was failing with a
cryptic error, because somewhere down the line some code was
interpredeted as an imcomplete string. This is in contrast to R's
ggplot2, wich just seems to ignore missing dots. A dot is now just added
silently if missing. Alternative an exception could be thrown, but I
think this is more userfriendly. At least it should not fail with a
cryptic error as before.