Fix axis ticks for MPL>=3.1.1
After https://github.com/matplotlib/matplotlib/pull/12839
axis ticks showed up along all the edges of the plot panel
and theming of tick labels was finicky.

The solution is we have to be careful in how we toggle the
visibility of these elements. If not careful, merely setting
the tick/tick labels could switch on the visibility. This
could interfere with theme.

closes #278