Fix: annotation_stripes with single stripe
annotation_stripes crashes when there is only a single break:

```python
       n_stripes = len(breaks)
        diff = np.diff(breaks)
>       step = diff[0]
E       IndexError: index 0 is out of bounds for axis 0 with size 0
geoms/annotation_stripes:82: IndexError
```

This let's it draw a single stripe instead.

There is a slight discrepancy in that multiple stripes
leave space at the sides, and the single one does not
(as if fill_range was set to True),
but I believe this is actually upstream of annotation_stripes,
when axis extends are calculated.