Fix skewed arrow head for facetted plots
When drawing an arrow, we use the relative dimensions of the plot
to compute an arrow head that is not skewed. The mistake was we
used the dimensions of the figure instead those of the axes. For
a single panel, the figure and axes have just about the same
relative dimensions (height/width). This is not the case when the
plot has many panels.

fixes #388