Add optional 'flat' parameter to geom_violin
This patch adds an optional 'flat' parameter to geom_violin which allows drawing of 'flat-violin' or 'half-violin' plots. This effectively removes the left half of the violin, drawing the density only on the right side of the sample position. Inspired by https://gist.github.com/dgrtwo/eb7750e74997891d7c20