Add optional 'style' parameter to geom_violin (#463)
* Add optional 'flat' parameter to geom_violin

This patch adds an optional 'flat' parameter to geom_violin which allows drawing of 'flat-violin' or 'half-violin' plots. This effectively removes the left half of the violin, drawing the density only on the right side of the sample position. Inspired by https://gist.github.com/dgrtwo/eb7750e74997891d7c20

* Add tests for flat violin plot

* Add missing paranthesis

There was a missing paranthesis in the added test for the 'flat' parameter causing a syntax error.

* Replaced 'flat', 'flat_side' parameter with 'style'

* Fix test image dimensions

* Fixed failing test if 'style' parameter isn't set

* Fix flat test image dimensions and dpi

* Rename 'flat' violin test to 'style'

Co-authored-by: Christian Meyenburg <meyenburg@zbh.uni-hamburg.de>