Fix FutureWarning for GeoDataFrame
In the future, it will not be possible to create an
active geometry column just by key assignment.

   data["geometry"] = other["geometry"]

This is how our aesthetic evaluation function has worked.
Solution is to create the evaluated frame in the
contructor once we have all the data. GeoDataFrames are
do take geometry columns this way.

The other option is

   other.set_geometry("geometry")

which returns a new dataframe. We cannot map to one column
this way.