Put tick pad with label for layout calculation
In matplotlib the space between the axis ticks and the axis
tick labels is associated with the ticks. In plotnine that space
is associated with the labels.

It is an edgecase, were when the ticks are blank and the labels are not
we must account for that space. e.g. Adding axis_text to theme_void.